requirements.txt
boto3==1.9.130
Django==2.1.7
django-heroku==0.3.1
django-cors-headers==2.5.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
httpie==1.0.2
psycopg2==2.7.7
python-dateutil==2.6.1
virtualenv==16.0.0
